resurrect upload request specs & fix response codes
