+++ spec/features/shared/authorization_error.rb
@@ -6,7 +6,7 @@ shared_examples "authorization error" do
   it "displays error" do
     visit path
     within(".modal") do
-      expect(page).to have_css(".modal-header", text: "Request ERROR 500")
+      expect(page).to have_css(".modal-header", text: "Request ERROR 403")
       expect(page).to have_css(".modal-body", text: "System-admin scope required")
     end
   end
